Refactor 2 Complete (#6)
* Generated code revert.
Reverted changes to the generated code app_event and InnerStackTrace (few weren't arrays where they should be in swagger) to their proper change before the extra swagger added fields.

* Fixed Comments.
Fixed spelling, and added an overview on how event_trace_builder works to the main method.

* Contructor and struct fixes.
Fixed the Contructor eventsApi initalization logic, changed the class to poin to the renamed eventsApi vs the old eventsAPI, removed the unnessisary URL Field, fixed the comment for the struct and spelling in it.

* Refactor 2

Changes to TrakerrClient to support loglevel and deploymentstage. Env name should now be working properly.

* Refactor 2

Changes to TrakerrClient to support loglevel EnvLanguage, and deploymentstage. Env name should now be working properly.